+++ test/eex/html_test.exs
@@ -3,21 +3,22 @@ defmodule EEx.HTMlTest do
   import EEx.HTML
   doctest EEx.HTML
 
-  test "escapes HTML" do
-    assert escape_to_binary("<script>") == "&lt;script&gt;"
-    assert escape_to_binary("html&company") == "html&amp;company"
-    assert escape_to_binary("\"quoted\"") == "&quot;quoted&quot;"
-    assert escape_to_binary("html's test") == "html&#39;s test"
+  test "raw accepts binary" do
+    assert "Hello, World!" == "#{raw("Hello, World!")}"
   end
 
-  test "escapes HTML to iodata" do
-    assert iodata_escape("<script>") == "&lt;script&gt;"
-    assert iodata_escape("html&company") == "html&amp;company"
-    assert iodata_escape("\"quoted\"") == "&quot;quoted&quot;"
-    assert iodata_escape("html's test") == "html&#39;s test"
+  test "raw accepts iolist" do
+    assert "Hello, World!" == "#{raw(["Hello, ", ["World!"]])}"
   end
 
-  defp iodata_escape(data) do
-    data |> escape_to_iodata() |> IO.iodata_to_binary()
+  test "raw accepts any term that implements String.Chars" do
+    assert "Hello, World!" == "#{raw(:"Hello, World!")}"
+    assert "125" == "#{raw(125)}"
+  end
+
+  test "raises ArgumentError when not an iolist" do
+    assert_raise(ArgumentError, fn ->
+      raw([:foo])
+    end)
   end
 end